chore: add scala-js support for fibonacci
